FINALLY WORKED AROUND THE DOUBLE QUOTES
I'll get to writing up an issue for this on *someone's* repo, not sure if it's a Contentful or `spike-contentful` issue.  Anyway, inline HTML in Contentful's markdown gets disgustingly broken in the final output (code missing, attributes doubly surrounded by quotes) unless you remove alllllll quotes from any attributes in your tags.  `<a href="https://google.com"></a>` should be corrected in Contentful to `<a href=https://google.com></a>`; either Contentful or `spike-contentful` are adding a second set of quotes to the HTML unless this correction is made.