Properly impl. base layout exception for homepage
I'm thinking it would be beneficial to Svelte beginners to have the docs
emphasize that you can have multiple <script> tags in your Svelte
components.  This is the case with valid HTML, yes, but coming from
Vue SFCs (the closest analogue to Svelte's IMO) I was under the
impression you could only have 1 of each tag.

Then again, it's been a few years since I've used Vue so I'm not 100%
sure this is relevant commentary 🤷‍♂️