Hey Sanity, it's not you, it's me...
Well honestly, it's you.

I gave it a good run, an honest try, *multiple times*, but too many
bumps in the road keep putting me off of Sanity. I'd get into it but
this is just a commit message lol, I'll likely write a blog post about
it though.

Long story short, I've returned to Ol' Reliable: DatoCMS! Probably
just going to stick with it from now on.

Anyway, the changes in this commit are as follows:
- Removed Sanity from the codebase, including npm packages and
  relevant code
- Added DatoCMS in, adjusted queries and edited components to
  accommodate, also added fetching for the blog (realized it lives
  in its own instance)
- Converted the whole thing to TS