refactor: We're doin it live!
Instead of leaving my site in indefinite, inaccessible limbo while this
overhaul is being worked on, I'm going to build this plane while
flying it, starting with the blog! Can't write about the experience in
real-time if there's nowhere to write 👉😎👉

All pages except the blog will be "turned off" and revived as they are
completed, which won't be until the "shell" of the site is complete.