refactor: Impl new gql clients
This required a move of all page.ts files to the server so the gql
clients could load their private env vars. It's a better decision for
client bundle size and perf, too; why ship what's not intended to run
after build time?

As a bonus, I applied the new indentation to the base +page.svelte.